fix: null the I2C config struct on declaration
Not nulling the I2C config on declaration can lead to errors like:
"i2c: i2c_param_config(760): i2c clock choice is
invalid, please check flag and frequency".
This is because the clk_flags member of the i2c_config_t struct was not
initialized, leading to an undefined value.